refactor: cartItems를 변경하는 함수를 최상단 컴포넌트에서 관리하도록 수정
