+++ src/components/CartItemList/CartItemList.tsx
@@ -1,31 +1,20 @@
-import React from "react";
 import { CartItemType } from "../../types/response";
 import CartItem from "../CartItem/CartItem";
 import { Container } from "./CartItemList.styles";
 
 interface CartItemListProps {
   cartItems: CartItemType[];
-  setCartItems: React.Dispatch<React.SetStateAction<CartItemType[]>>;
+  updateCartItem: (cartId: number, cartItem: CartItemType) => void;
   isSelected: (cartId: number) => boolean;
   toggleSelect: (cartId: number) => void;
 }
 
 function CartItemList({
   cartItems,
-  setCartItems,
+  updateCartItem,
   isSelected,
   toggleSelect,
 }: CartItemListProps) {
-  const handleCartItem = (cartId: number, newItem: CartItemType) => {
-    setCartItems((prevItems) => {
-      const targetIndex = prevItems.findIndex((item) => item.id === cartId);
-      const newCartItems = [...prevItems];
-      newCartItems[targetIndex] = newItem;
-
-      return newCartItems;
-    });
-  };
-
   return (
     <section css={Container}>
       {cartItems.map((cartItem) => {
@@ -33,7 +22,7 @@ function CartItemList({
           <CartItem
             key={cartItem.id}
             cartItem={cartItem}
-            handleCartItem={handleCartItem}
+            updateCartItem={updateCartItem}
             isSelected={isSelected(cartItem.id)}
             toggleSelect={() => toggleSelect(cartItem.id)}
           />
